+++ core/index/sorted_column.cpp
@@ -28,7 +28,7 @@
 
 namespace irs {
 
-bool sorted_column::flush_sprase_primary(
+bool sorted_column::flush_sparse_primary(
   doc_map& docmap, const columnstore_writer::values_writer_f& writer,
   doc_id_t docs_count, const comparer& compare) {
   auto comparer = [&compare, this](
@@ -51,10 +51,10 @@ bool sorted_column::flush_sprase_primary(
 
   std::vector<size_t> sorted_index(index_.size() - 1);
   std::iota(sorted_index.begin(), sorted_index.end(), 0);
-  std::stable_sort(sorted_index.begin(), sorted_index.end(),
-                   [&comparer, this](size_t lhs, size_t rhs) {
-                     return comparer(index_[lhs], index_[rhs]);
-                   });
+  std::sort(sorted_index.begin(), sorted_index.end(),
+            [&comparer, this](size_t lhs, size_t rhs) {
+              return comparer(index_[lhs], index_[rhs]);
+            });
 
   doc_id_t new_doc = doc_limits::min();
 
@@ -107,7 +107,7 @@ std::pair<doc_map, field_id> sorted_column::flush(
   auto [column_id, column_writer] =
     writer.push_column(info_, std::move(finalizer));
 
-  if (!flush_sprase_primary(docmap, column_writer, docs_count, less)) {
+  if (!flush_sparse_primary(docmap, column_writer, docs_count, less)) {
     flush_already_sorted(column_writer);
   }
