+++ microbench/CMakeLists.txt
@@ -20,6 +20,7 @@ endif ()
 
 add_executable(iresearch-microbench
   ./top_term_collector_benchmark.cpp
+  ./hash_map_benchmark.cpp
   ./segmentation_stream_benchmark.cpp
   ./simd_utils_benchmark.cpp
   ./lower_bound_benchmark.cpp
